Refactor unittest script
This way I can run it locally without it archiving results to the master

+++ buildbot/unittest.sh
@@ -1,10 +1,5 @@
 #!/bin/bash
 
-if [ -z "$BUILDER_NAME" -o -z "$BUILD_NUMBER" ] ; then
-    echo "Must specify BUILDER_NAME and BUILD_NUMBER"
-    exit 1
-fi
-
 # Run tests
 APK="project/build/apk/project-Expedia-debug-unaligned.apk"
 TEST_APK="project/build/apk/project-Expedia-debug-test-unaligned.apk"
@@ -21,13 +16,15 @@ java \
 
 SPOON_RESULT=$?
 
-# Archive to the master
-tar cvzf "$OUTPUT_TAR" 'spoon/unit'
-scp "$OUTPUT_TAR" "buildbot@buildbot.mobiata.com:/home/buildbot/artifacts/."
-ssh "buildbot@buildbot.mobiata.com" 'cd /home/buildbot/artifacts ; umask 022 ; for i in *.tar.gz ; do echo Extrating "$i" ; tar xzf "$i" ; rm -f "$i" ; done'
+if [ -n "$BUILDER_NAME" -a -n "$BUILD_NUMBER" ] ; then
+    # Archive to the master
+    tar cvzf "$OUTPUT_TAR" 'spoon/unit'
+    scp "$OUTPUT_TAR" "buildbot@buildbot.mobiata.com:/home/buildbot/artifacts/."
+    ssh "buildbot@buildbot.mobiata.com" 'cd /home/buildbot/artifacts ; umask 022 ; for i in *.tar.gz ; do echo Extrating "$i" ; tar xzf "$i" ; rm -f "$i" ; done'
 
-## Cleanup locally
-rm -rf "spoon"
-rm -f spoon-unit-*.tar.gz
+    ## Cleanup locally
+    rm -rf "spoon"
+    rm -f spoon-unit-*.tar.gz
+fi
 
 exit "$SPOON_RESULT"
